DB-9843 fix rebase issues, add unit test getCsvOptions
