+++ hbase_sql/src/test/java/com/splicemachine/derby/impl/sql/execute/operations/ExternalTableUnitTests.java
@@ -1,13 +1,31 @@
+/*
+ * Copyright (c) 2012 - 2020 Splice Machine, Inc.
+ *
+ * This file is part of Splice Machine.
+ * Splice Machine is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU Affero General Public License as published by the Free Software Foundation, either
+ * version 3, or (at your option) any later version.
+ * Splice Machine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero General Public License for more details.
+ * You should have received a copy of the GNU Affero General Public License along with Splice Machine.
+ * If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
 package com.splicemachine.derby.impl.sql.execute.operations;
 
+import com.splicemachine.db.iapi.error.StandardException;
 import com.splicemachine.derby.stream.spark.SparkExternalTableUtil;
 import com.splicemachine.derby.stream.utils.ExternalTableUtils;
+import com.splicemachine.system.CsvOptions;
 import org.apache.hadoop.fs.Path;
 import org.apache.spark.sql.types.DataTypes;
 import org.apache.spark.sql.types.StructType;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.io.*;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
@@ -171,5 +189,26 @@ public class ExternalTableUnitTests {
                         "PARTITIONED BY(col1, col2);",
                 ExternalTableUtils.getSuggestedSchema(s, part).toString() );
 
+    }
+
+    @Test
+    public void testGetCsvOptions() throws StandardException, IOException {
+        CsvOptions opt = new CsvOptions("#", "!", "\n");
+
+        String expected = "{lineSep=\n" +
+                ", timestampFormat=yyyy-MM-dd'T'HH:mm:ss.SSSZZ, escape=!, sep=#}";
+        Assert.assertEquals(expected, SparkExternalTableUtil.getCsvOptions(opt).toString());
+
+        // test serde
+        ByteArrayOutputStream stream = new ByteArrayOutputStream();
+        ObjectOutputStream oos = new ObjectOutputStream(stream);
+        opt.writeExternal(oos);
+        oos.flush();
+
+        ByteArrayInputStream in = new ByteArrayInputStream(stream.toByteArray());
+        ObjectInputStream ois = new ObjectInputStream(in);
+        CsvOptions opt2 = new CsvOptions(ois);
+
+        Assert.assertEquals(expected, SparkExternalTableUtil.getCsvOptions(opt2).toString());
     }
 }
